+++ drift/test/integration_tests/migrations_integration_test.dart
@@ -534,6 +534,19 @@ void main() {
 
     expect(underlying.userVersion, 3);
   });
+
+  test("alterTable works for databases that can't set legacy alter table",
+      () async {
+    final interceptor = _NoLegacyAlterTable();
+    final db = TodoDb(NativeDatabase.memory().interceptWith(interceptor));
+    addTearDown(db.close);
+
+    final user = await db.users.insertReturning(
+        UsersCompanion.insert(name: 'test user', profilePicture: Uint8List(0)));
+    await Migrator(db).alterTable(TableMigration(db.users));
+    expect(await db.users.all().get(), [user]);
+    expect(interceptor.didPreventLegacyAlterTable, isTrue);
+  });
 }
 
 class _TestDatabase extends GeneratedDatabase {
@@ -547,4 +560,19 @@ class _TestDatabase extends GeneratedDatabase {
 
   @override
   final MigrationStrategy migration;
+}
+
+class _NoLegacyAlterTable extends QueryInterceptor {
+  var didPreventLegacyAlterTable = false;
+
+  @override
+  Future<void> runCustom(
+      QueryExecutor executor, String statement, List<Object?> args) {
+    if (statement.contains('legacy_alter_table') && statement.contains('=')) {
+      didPreventLegacyAlterTable = true;
+      throw 'not allowed';
+    }
+
+    return super.runCustom(executor, statement, args);
+  }
 }